Create Wiki-Text-based language models for TextCat
Moved existing query-based models to LM-query/. Created 70 new models
based on random articles from the relevant Wikipedia. Minor updates
to PHP code, including change output join text to "OR" so as not to
conflict with language model "or.lm". Major updates to README.md.

These models have not been evaluated (see T121539), but are made
available as is.

BUG: T121545

Change-Id: I772670f2fa97dfe3981fd139ea40c62f921ccda7